Rename druid-win-shell to druid-shell
As part of platform independence, rename druid-win-shell to remove the
reference to a specific platform.

Also fixes druid-shell examples, which were slightly broken.

Normalizes line endings to LF (not CRLF).

Part of #13
