+++ druid-shell/src/application.rs
@@ -1,0 +1,9 @@
+use win_main;
+
+pub struct Application;
+
+impl Application {
+    pub fn quit() {
+        win_main::request_quit();
+    }
+}
