+++ druid-shell/src/dcomp.rs
