--- druid-win-shell/src/application.rs
@@ -1,9 +1,0 @@
-use win_main;
-
-pub struct Application;
-
-impl Application {
-    pub fn quit() {
-        win_main::request_quit();
-    }
-}
