+++ examples/anim.rs
@@ -15,15 +15,15 @@
 //! Example of animation frames.
 
 extern crate druid;
-extern crate druid_win_shell;
+extern crate druid_shell;
 extern crate kurbo;
 extern crate piet;
 
 use kurbo::Line;
 use piet::RenderContext;
 
-use druid_win_shell::win_main;
-use druid_win_shell::windows::WindowBuilder;
+use druid_shell::win_main;
+use druid_shell::windows::WindowBuilder;
 
 use druid::{Ui, UiMain, UiState};
 
@@ -90,7 +90,7 @@ impl AnimWidget {
 }
 
 fn main() {
-    druid_win_shell::init();
+    druid_shell::init();
 
     let mut run_loop = win_main::RunLoop::new();
     let mut builder = WindowBuilder::new();
