+++ examples/dynamic.rs
@@ -15,12 +15,12 @@
 //! An example of dynamic graph mutation.
 
 extern crate druid;
-extern crate druid_win_shell;
+extern crate druid_shell;
 
 use std::collections::BTreeMap;
 
-use druid_win_shell::win_main;
-use druid_win_shell::windows::WindowBuilder;
+use druid_shell::win_main;
+use druid_shell::windows::WindowBuilder;
 
 use druid::widget::{Button, Column, EventForwarder, Label, Padding, Row};
 use druid::{Id, UiMain, UiState};
@@ -41,7 +41,7 @@ enum Action {
 }
 
 fn main() {
-    druid_win_shell::init();
+    druid_shell::init();
 
     let mut run_loop = win_main::RunLoop::new();
     let mut builder = WindowBuilder::new();
