+++ examples/ext_event.rs
@@ -15,18 +15,18 @@
 //! Example of sending external events to the UI.
 
 extern crate druid;
-extern crate druid_win_shell;
+extern crate druid_shell;
 
 use std::{thread, time};
 
-use druid_win_shell::win_main;
-use druid_win_shell::windows::WindowBuilder;
+use druid_shell::win_main;
+use druid_shell::windows::WindowBuilder;
 
 use druid::widget::Label;
 use druid::{UiMain, UiState};
 
 fn main() {
-    druid_win_shell::init();
+    druid_shell::init();
 
     let mut run_loop = win_main::RunLoop::new();
     let mut builder = WindowBuilder::new();
